Workspace/requirements.txt
Flask
Pillow
stegano==0.11.3
pycryptodome
Werkzeug